Made some fixes
In LUA we don't need to make the queryData's a string. {} is a valid
"table" definition. We are already taking the queryData in db.lua and
json_encoding it that converts the LUA variable to a json string at that
point.

Additionally there is a way to do inserts or updates...I looked back at
the essentialmode code again now that I'm understanding this more and I
found that you will typically want to do a POST. For a column that we
create maybe "identifier". We need to assign this column which is
basically a players steam id and is separate from the _id and _rev. If
the POST returns a nil responseText then we need to generate a _id to
use in the identifier parameter for PUT. So do a GET to "_uuid" and in
the callback grab the _uuid by doing rText.uuids[1]. Then use that
generated uuid to do a PUT since there is no record that matches the
players identifier column. If there is a match ... basically the else
from the if responseText == nil. Then we want to grab the _id from the
returned response. Like so: responseText._id and use that as the
identifier paramater to PUT. The identifier parameter is NOT the same as
the player's target.identifier this is the _id or _uuid. Then we want to
grab the _rev from the POST and fill that in inside the queryData... and
set a column called "identifier" as the players target.identifier. So
next time we do a POST there will be a record that matches the
identifier for that player and will instead update the _id instead of
generate a new one.