Added to main repo again.
From my understanding it was okay to add it to my main repo (having the fork repo on my PC under my main repo). Will push to both this repo and fork repo with any changes made and keep history. If there is a problem with the fs_freeroam group, then please let me know.